requirements.txt
torch
flask
diffusers
transformers
accelerate
Pillow
